Handle disabled watchdog (WATCHDOG_USEC not set)
Systemd will not set the `WATCHDOG_USEC` environment variable if the
unit in question has the watchdog disabled by setting `WatchdogSec=0`.

Before this commit, healthdog would fail in this case, exiting before it
even started the process it was supposed to run and watch.

With this commit, if `WATCHDOG_USEC` is not set, healthdog will run
without errors, will start the desired process, and will refrain from
petting Systemd's watchdog.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: minor