Make sure the command to be monitored is passed
The command to execute (that is, the one that is meant to be monitored
by the health checks) is a requited argument, but previously this was
not checked. (In fact, failing to pass the command would cause a panic.)

This commit checks if the command is passed and updates the CLI help
accordingly.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>